Adjust short shield duration
Right now the short shield is way too long, so taking a full bomb explosion only deals 3 1/2 hearts of damage (in the original game the damage is like double that or more). Reducing the short shield duration so that it deals 6 full hearts, which is closer to the original game.