use std::filesystem::create_directories in lif::createDirIfNotExisting
This automatically handles the creation of the whole directory tree
instead of failing if the parent directories don't exist, and it works
on all platforms the same way.
The downside is that we now need C++17 to build the game, but at this
point it's probably safe to assume it's supported on all the platforms
we care about.