Replace nlohmann::json with tinyjson
Our version of nlohmann::json doesn't compile any more with C++20 so
we should update it if we want lifish to keep working in the future.
However nlohmann::json is now a bloated beast of over 25000 lines of
code (more than twice our current version).
This is a ridiculous amount of code to carry around just to parse a
few JSON files. We don't need most of the features nlohmann::json
provides and we don't particularly need a lot of speed, as our files
are so small that they are still fast to parse even with a less optimized
library.

For now I'm pulling in a slightly modified version of Tinyjson, which
has the huge advantage of being 1/25th the size of nlohmann while still
providing everything we need.
In the future, if for any reason it's needed, we might replace it with
a handmade json parser. Or at that point, might as well use a custom
format. JSON is probably not what I'd have chosen today, but eh, it is
what it is.

NOTE: this was not *extensively* tested, we probably want some more
testing before the next release